Init impl of IGPostView widget
