+++ lib/pages/test_instagram.dart
@@ -3,7 +3,7 @@ import 'dart:convert' as JSON;
 import 'dart:core';
 import 'package:kronogram/services/database.dart';
 import 'package:kronogram/services/authentication.dart';
-import 'package:kronogram/models/KronoInstaPost.dart';
+import 'package:kronogram/models/InstaPost.dart';
 import 'package:http/http.dart' as http;
 
 class InstagramPage extends StatefulWidget {
@@ -20,7 +20,7 @@ class InstagramPage extends StatefulWidget {
 }
 
 class _InstagramPageState extends State<InstagramPage> {
-  List<KronoInstaPost> posts = new List();
+  List<InstaPost> posts = new List();
 
   Future getPosts() async {
     var instaUser = await widget.db.getInstagramInfo(widget.userId);
@@ -30,7 +30,7 @@ class _InstagramPageState extends State<InstagramPage> {
     final jsonPostHistory = JSON.jsonDecode(graphResponse.body);
     print(jsonPostHistory);
     for (var jsonPost in jsonPostHistory['data']) {
-      KronoInstaPost instaPost = new KronoInstaPost.fromJson(jsonPost);
+      InstaPost instaPost = new InstaPost.fromJson(jsonPost);
       if (jsonPost['media_type'] == 'CAROUSEL_ALBUM') {
         var id = instaPost.getID();
         var albumResponse = await http.get(
