+++ pubspec.yaml
@@ -38,6 +38,7 @@ dependencies:
   flutter_svg: ^0.17.1
   twitter_api: ^0.1.2
   tweet_ui: ^1.1.0
+  cached_network_image: ^2.0.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
